+++ ng-workspace/projects/angulardev/src/app/app.component.ts
@@ -22,7 +22,8 @@ export class AppComponent {
 
 
   /**
-   * Constructor for the Angular Dev App Component.
+   * Constructor for the Angular Dev App Component. The version number below is
+   * used by the `create-common-package.bash` file to determine the app version.
    * Version: 3.0.0
    * @param owfService 
    * @param document 
@@ -35,11 +36,9 @@ export class AppComponent {
    * 
    */
   ngOnInit(): void {
-    // Called after the constructor, initializing input properties, and the first call to ngOnChanges.
-    // Set the app's favicon.
+    // Called after the constructor, initializing input properties, and the first
+    // call to ngOnChanges. Set the app's favicon.
     this.document.getElementById('appFavicon').setAttribute('href', 'assets/app/img/OWF-Logo-Favicon-32x32.ico');
-    
-
   }
 
   /**